02-backend/plausibility_notes/pyproject.toml
[project]
name = "plausibility-notes"
version = "0.1.0"
description = "Plausibility Notes Microservice"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "prometheus-fastapi-instrumentator",
    "typing-extensions>=4.5.0",
    "fastapi>=0.120.1",
    "uvicorn>=0.41.0",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.1.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "h11>=0.16.0",
    "jaraco.context>=6.1.0",
    "event-logging",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1,<9",
    "pytest-asyncio>=1.1.0,<2",
    "httpx>=0.28.1,<0.29",
    "pytest-cov>=6.2.1,<7",
    "interrogate>=1.7.0,<2",
    "pytest-env>=1.1.5,<2",
    "pytest-httpx>=0.35.0,<0.36",
    "pre-commit>=4.3.0,<5",
    "aiosqlite>=0.20.0,<0.22",
    "mypy>=1.19.1,<2",
    "ruff>=0.14.6,<0.15",
    "pyrefly>=0.49.0",
]

[tool.uv]
package = false

[tool.uv.sources]
event-logging = { workspace = true }


[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
markers = [
    "integration: mark a test as an integration test (uses real external services)"
]
addopts = "-m 'not integration' --cov=src --cov-report=term-missing"
testpaths = [
    "tests"
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["src/models/db/migrations/**"]

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-module = true
ignore-nested-functions = true

[tool.pyrefly]
search_path = ["."]
project-includes = ["."]
disable-search-path-heuristics = true

[tool.pyrefly.errors]
implicit-any = "error"
unannotated-parameter = "error"
unannotated-return = "error"
unannotated-attribute = "error"
implicit-abstract-class = "error"
implicitly-defined-attribute = "error"
missing-override-decorator = "error"
not-required-key-access = "error"
open-unpacking = "error"
untyped-import = "error"
missing-source = "error"
missing-source-for-stubs = "error"
deprecated = "error"
redundant-cast = "error"
unnecessary-comparison = "error"
